Fix many compiler warnings, and a few things warned by clang-tidy and cppcheck.  This has changed a few interfaces!
Most usage not break but small updates may be needed to user code.   (Examples of interface changes are e.g. use const references for objects, some parameters that should have been size_t or unsigned have been changed from ints, deprecate unsafe calls and add alternative safer call, etc.)

+++ include/Aria/ArArgumentBuilder.h
@@ -215,8 +215,8 @@ protected:
    * with space-quote, then the argument must end with a quote-space.
   **/
   bool isEndArg(const char *buf, 
-                int len, 
-                int &indexInOut,
+                size_t len, 
+                size_t &indexInOut,
                 int endArgFlags);
 
   /// Determines whether the specified character is an acceptable space (either normal or extra)
